requirements.txt
pandas
pyarrow
pyyaml
matplotlib
streamlit